little code cleanup
